backend_agent/requirements.txt
fastapi
uvicorn[standard]
openai
asyncpg
httpx
python-dotenv
pydantic
